deprecate  casm! in prover unit tests
